pyproject.toml
[tool.poetry]
name = "snakemake-executor-plugin-googlebatch"
version = "0.1.0"
description = ""
authors = [
    "Vanessa Sochat <sochat1@llnl.gov>",
    "Johannes Koester <johannes.koester@uni-due.de>"
]
readme = "README.md"
license = "MIT"
repository = "https://github.com/snakemake/snakemake-executor-plugin-googlebatch"
documentation = "https://snakemake.github.io/snakemake-plugin-catalog/plugins/executor/googlebatch.html"
keywords = ["snakemake", "plugin", "executor", "cloud", "google-batch"]

[tool.poetry.dependencies]
python = "^3.11"
google-cloud-batch = "^0.17.1"
requests = "^2.31.0"
google-api-core = "^2.12.0"
google-cloud-storage = "^2.12.0"
snakemake-interface-common = "^1.14.0"
snakemake-interface-executor-plugins = "^8.0.2"
jinja2 = "^3.1.2"
google-cloud-logging = "^3.8.0"

[tool.poetry.group.dev.dependencies]
black = "^23.9.1"
flake8 = "^6.1.0"
coverage = "^7.3.1"
pytest = "^7.4.2"
snakemake = {git = "https://github.com/snakemake/snakemake.git", branch="main"}
snakemake-storage-plugin-s3 = "^0.2.5"

[tool.coverage.run]
omit = [".*", "*/site-packages/*", "Snakefile"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"